backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.9
httpx==0.26.0
python-dotenv==1.0.1
python-dateutil==2.8.2
